Remove hasattr
